Refactored the ManGui.py to fit our folder structure
